+++ lib/serega/plan.rb
@@ -69,10 +69,6 @@ class Serega
     # SeregaPlan instance methods
     #
     module InstanceMethods
-      # Shows if plan includes batch points
-      # @return [Array<SeregaPlanPoint>] points to serialize
-      attr_reader :has_batch_points
-
       # Parent plan point
       # @return [SeregaPlanPoint, nil]
       attr_reader :parent_plan_point
@@ -100,7 +96,6 @@ class Serega
       # @return [SeregaPlan] Serialization plan
       #
       def initialize(parent_plan_point, modifiers)
-        @has_batch_points = false # should be before assigning points, generated points can change this attribute
         @parent_plan_point = parent_plan_point
         @points = attributes_points(modifiers)
         @points_hash = points.to_h { |point| [point.name, point] }
@@ -120,17 +115,6 @@ class Serega
       #
       def data_class
         @data_class ||= self.class.data_class_for(point_names)
-      end
-
-      #
-      # Marks current plan and top-level plans as `has_batch_points`
-      # It is needed to initialize Batch Attribute Loaders when serialization starts
-      #
-      def mark_as_has_batch_points
-        return if has_batch_points
-
-        @has_batch_points = true
-        parent_plan_point.plan.mark_as_has_batch_points if parent_plan_point # rubocop:disable Style/SafeNavigation
       end
 
       private